% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifespan.R
\name{lifespan}
\alias{lifespan}
\title{lifespan}
\usage{
lifespan(nx)
}
\arguments{
\item{nx}{A Matrix population model or an Integral projection model square matrix with growth, survival, and fecundity (in the top row only) as transitions in each kernel or cell}
}
\description{
Take a projection matrix, a population of 100 seedlings (recruits), remove the reproductive transitions (first row all zero) and measure years until the 100 individuals are dead
}
