% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeedHarvestSim.R
\name{SeedHarvestSim}
\alias{SeedHarvestSim}
\title{Seed Harvest Simulation}
\usage{
SeedHarvestSim(
  Mx_list,
  TMx_list,
  Nx,
  StartPopSize,
  GoodBadTm,
  Freq = c(0, 0),
  Int = c(0, 0),
  TotYrs,
  ClusteredColl = "N",
  seedbank = FALSE
)
}
\arguments{
\item{Mx_list}{A list of annual transition matrices indicating survival and growth from col to row; fecundity in the first row (square matrix with N stages).}

\item{TMx_list}{A list of annual transition matrices without fecundity (square matrix with N stages).}

\item{Nx}{A population size vector by stage (length N). Stable stage distriubiton or most recent count by stage.}

\item{StartPopSize}{A number for the starting population size of the simulation.}

\item{GoodBadTm}{A transition matrix (row to col) of good and bad years. Annual transition matrices are defined as bad < mean fecundity <= good}

\item{Freq}{A numerical vector for the the probability of collecting seed given the year c(good, bad).}

\item{Int}{A numerical vector for the intensity (the percentage decrease in fecundity) if seed harvest for c(good, bad).}

\item{TotYrs}{A number indicating how many years to project in the simulation.}

\item{ClusteredColl}{"Y" or "N" indicating if collection years should be clustered. Results in increased likelihood of collection in t+1 when collecting in year t}

\item{seedbank}{TRUE or FALSE to skip the first element in the first row when reducing fecundity.}
}
\value{

}
\description{
Simulation of climate and seed harvest frequency and intensity
}
